complete directory restructure for library development
