+++ tsconfig.spec.json
@@ -1,0 +1,5 @@
+{
+    "compilerOptions": {
+      "types": ["jest", "node"]
+    }
+  }
